// types/demo.d.ts
declare namespace Demo {
    // Interfaces
    interface Base {
        name: string;
        icon: Icon;
        objectURL?: string;
    }

    interface IFile extends Base {
        date: string;
        fileSize: string;
    }

    interface Metric {
        title: string;
        icon: string;
        color_light: string;
        color_dark: string;
        textContent: MetricContent[];
        color?: string;
        fieldColor?: string;
        files?: string;
        fileSize?: string;
    }
// ...
    interface MetricContent {
        amount: string;
        text: string;
    }
// ...
}